fix(Menu, MenuLayer, Cookies, LegalTerms): Avoid displaying two h1 at the same time
Add a new prop to Menu, the component that renders conditionally a default h1 when needed. Now we control from each view: LegalTerms and Cookies if this default title is needed, since it is not by default because they provide a h1, but we have to remove it when CoursesList or Contact is open.

In order to achieve that we intercept the handler to update the state of an inner prop and need to add a timeout before removing the page h1 to make it always visible untill the new h1 is visible"